requirements.txt
customtkinter>=5.2.0
simpleaudio>=1.0.4
Pillow>=9.0.0
